Update bean, new test web service
Update bean class, modify estimating-config, create new test webservice

+++ estimating-v3/src/main/java/com/estimating/dao/core/AbstractBaseDao.java
@@ -72,7 +72,7 @@ public abstract class AbstractBaseDao<T> implements IBaseDao<T> {
 	 * @param primaryKey
 	 * @return only entity from parameter
 	 */
-	public T findOneById(Long primaryKey) {
+	public T findOneById(int primaryKey) {
 		return this.entityManager.find(entityClass, primaryKey);
 	}
